chat-with-college-notes/requirements.txt
streamlit
google-genai
pypdf
numpy
